# Pluggable certificate storage (following on from #284)